Cleaner solution...
The two methods that differ get their own class. It's likely the smallest
fully functional C++ class I ever wrote with an actual code of about
10 lines, but it does exactly what is not generic between the two server
types.

The new MSNAuthSocket contains the rest and will be the base class for
MSNNotifySocket. MSNSocket will be used as 'raw' socket without negotiation
by MSNSwitchBoardSocket, but that's not planned for the near future.

MSNNotifySocket will be first, with everything else in place now maybe even
this weekend...

svn path=/trunk/kdenetwork/kopete/; revision=157250