CCMAIL=kopete-devel@kde.org
Finally made KopeteEmailWindow not suck! Whoohoo! All it needs now is
layout support. I think it's now at the stage where I can say that it
Works, and is safe to deploy. All it needs now is a *cough* CAPABILITIES
SYSTEM (hi Martijn!), to determine whether we can put the subject field
in or not.

So anyway, plugin authors, please start using this in your plugins. Jabber
already uses it, and has a configurable bit, asking whether the user wants
to do KCW or KEW by default, but they can explicitly select which. There's
an actual difference in the message type for Jabber with this, so other
plugins can probably ask what to do by default and that's it. The reply
button works, and so does the show next queueing stuff.

I think it would be nice if all plugins could offer users the choice between
KCW and KEW.

Cheers!
:) d

svn path=/trunk/kdenetwork/kopete/; revision=162641