Huuuuuuuuge commit: introducing KopeteMetaContact.
Far from finished, but what's there works, and it doesn't hurt existing code

MSN uses it. There are a lot of things I cannot do before all plugins are
ported, like breaking downwards compatibility, but I first want only a single
additional plugin over and incorporate any suggestions from that.

Reports about missing and unimplemented stuff and the ugly GUI *will* be
ignored for now, because this is work-in-progress.

OTOH, patches to implement the various FIXMEs and suggestions about the new
API are more than welcome.

svn path=/trunk/kdenetwork/kopete/; revision=162754