Huuuuuge API cleanup:
- Again zillions of cosmetic changes
- KopeteContact now *requires* a KopeteMetaContact as parent
- KopeteContactList::createContact is pure virtual, don't implement it!
- KopeteContactList::createContact no longer takes the ID as string,
  the serializedData is for that
- Fix MSN and ICQ to build against this improved API
- Several smaller changes

TODO is still a lot of work.

Most urgent is to provide proper serialized data to the plugins, so the
creation of KopeteContacts actually seems to work.

That, in turn, requires a proper serialize()/deserialize() call so the
data can be stored in the contact list.

At least it's less broken by design than it was, and it didn't actually
run anyway, so there's no regression either.

Any takers to pick up where I left? It's bed time here...

Ryan? :) :) :) :)

svn path=/trunk/kdenetwork/kopete/; revision=168257