Since there's already QObject::name it's not that smart to introduce a KopeteContact::name as well.
s/name/displayName/, but unfortunately not an automated find and replace
because name() was even defined virtual and there's some usage of
QColor::name(), which obviously shouldn't be replaced.

Needless to say this manual fixing sucks, and I'm quite sure I missed
some occasions. If contacts have the wrong name in the user list you know
what to look for now...

svn path=/trunk/kdenetwork/kopete/; revision=169151