Cache the online status and don't recalc it for each status() call again.
Call updateStatus() to update the online status.

Reason for this change is not really the performance boost (it's not that
slow without this either), but to be able to EASILY suppress emitting
contactStatusChanged signals when the status is unchanged, and to avoid
code duplication in the 4 methods that affect the online status.

This should close bug 47953, but I can't test before tomorrow.

Till: could you test and close the bug when appropriate? Thanks.

CCMAIL: Till Gerken <till@tantalo.net>

svn path=/trunk/kdenetwork/kopete/; revision=178812