Replace one compile breakage with another one ;-)
libkopete builds now, but the plugins need updating. Doning that as we speak.

This commit completes what I accidentally committed partly in my previous
commit: let libkopete determine the file name for logging and not the plugin

Remaining TODO-item, but less in a hurry:

'view history' should be moved outside the plugins or at least use the
updated path names.

svn path=/trunk/kdenetwork/kopete/; revision=192672