KJanusWidget doesn't like it if its child pages are deleted from ~QObject if that QObject was once known as 'kapplication'. It wants to update pages, which triggers QPaintDevice code, which doesn't work because ~QApp has closed the X server connection already.
Although this IMHO should be fixed outside Kopete this workaround avoids
the crashes. We simply makes sure to delete the prefs dialog before the
QApplication destructor is called, so the JanusWidget can still do as it
pleases.

(Analysis of the problem mailed to Simon, it's beyond my knowledge to dive
much deeper into it.)

svn path=/trunk/kdenetwork/kopete/; revision=194358