Use KStaticDeleter to delete the lib loader before ~kapp, not after.
Unfortunately this triggers another crash on exit because the deletion
order is once more different. According to valgrind I'm now doubly
deleting the plugins themselves. :(

*grmbl*

I consider the KSD the proper way though, and since we're not planning a
release within a few days I still commit this. Just don't exit Kopete and
everything is fine ;-)

svn path=/trunk/kdenetwork/kopete/; revision=194463