Fix some obscure problems in the DNS lookups:
// Ideally we want to the full connection to MSN to be handled async, but due
// to some design issues in QDns this fails if people with dialup connections
// start Kopete before their internet connection. The workaround from
// TrollTech is to not use QDns, but use the libc gethostbyname call instead.
// The sync calls in KExtendedSocket use this, only the async lookup uses DNS.
// This is slightly annoying as it blocks the GUI for the duration of the DNS
// lookup, but properly configured systems will hardly notice that. Besides,
// there's nothing we can do about it...  For Qt 4/KDE 4 we can hopefully
// leave the lookup to the socket again and remove the manual lookup call.
// This cannot be fixed in Qt 3 unfortunately.

This commit fixes the Kopete side of bug 50279, but in fact it's more a
KExtendedSocket/QDns problem (bug is already reassigned there), so I'm not
closing it.

It also works around the problem with recent glibcs for debian unstable
and Mandrake Cooker that breaks QDns, so it also 'fixes' bug 52494. Here
the real problem is beyond our reach as well, but it works nevertheless
as workaround for now.

CCMAIL: 50279@bugs.kde.org
CCMAIL: 52494-done@bugs.kde.org

svn path=/trunk/kdenetwork/kopete/; revision=196873