CCMAIL: kopete-devel@kde.org
Hm, need a cool name for this compromise, a la the Iceland Compromise[1].

How about the Switzerland Compromise? People will look back on the ages
and say "Oh yes, the Switzerland Compromise of 2003, that was a huge
seachange in the field of instant messanging." ;)

Reinstating the AIM directory until I can provide a TOC->OSCAR upgrade path
via kconf_update.

[1]: http://marc.theaimsgroup.com/?l=kde-cvs&m=101998422704131&w=2

svn path=/trunk/kdenetwork/kopete/; revision=197067