CCMAIL: kopete-devel@kde.org, jaymz@artificial-stupidity.net, nevyn@artificial-stupidity.net, klingens@kde.org, neil@kde.org Removing TOC once for all. To quote "Bill Lawry" in the 12th Man - The Last Dig: "Gone, yes, got 'im! Fuck off, you're out!".
I was barking up the wrong bush[1] when I checked for a migration path
the first time: as soon as I saw Nick's mail, it was blindingly obvious
what was done, and I felt stupid.

So, I'm removing TOC, which is deprecated, useless, and just plain shit,
in favour of OSCAR, which works and gains features, not loses them. Yay!

Enjoy,
:) d

[1]: Yes, that was deliberate.

svn path=/trunk/kdenetwork/kopete/; revision=197106