* if outgoing encrypted message is not in the cache, don't display   the ciphertext, but decrypt the message * therefore, always encrypt outgoing messages to self, too.
This hides another problem: for some reason, the cache is not working
properly. it does not find the ciphertext message and thus displays
the cipertext in the chatwindow instead of the cached plaintext.
looking at the code, I see no reason why it should not work, still...

Someone have a look here, please.

svn path=/trunk/kdenetwork/kopete/; revision=200047