Commit this monster, as it has grown waaaay too big for a local checkout...
Huge change in the handling of the contact list loading and saving.

The most important changes are:

- loading and saving is based on key/value pairs now for all data, and
  not anonymous string lists

- plugin data is prepared for storage in libkabc, including the actual
  data format as discussed with Neil. This means basically that the
  ids for multiple accounts in a single contact are joined using a
  standardized character, in this case QChar( 0xE000 ), which is reserved
  for private use like this and non-printable. Also, the protocol names
  are standardized, so jabber is stored as xmpp, aim/oscar and aim/toc
  are both stored as aim, etc. There is no actual kabc code yet, but
  with this commit everything needed is in place now to make the jump.

- contacts from aim/toc are now converted to aim/oscar so people don't
  need to recreate contacts. I know they are on the server too, but this
  way the mapping to a meta contact is no longer lost.

- lots of minor other changes as well

Since all this required some intrusive changes in the contactlist.xml
format I had to write a method for converting old-style files to the
new format. If it wasn't for this method this commit would have made the
total Kopete code some 300 lines smaller, now it makes it even bigger :(

On the positive side I can mention that tons of duplicated methods in the
various plugins are moved to a single location in libkopete, and moreover,
the rather error-prone loop for serializing multiple contacts in a single
meta contact is moved to libkopete, so plugins don't need to do all that
processing and tracking themselves.

TODO: Some plugins are not converted yet. People using SMS, WinPopup, Gadu
and the non-protocol plugins can better not CVS up yet, since the data
for these plugins will be lost. All other plugins are converted, though
given the massive size of this commit I'd be surprised if there are no
bugs. It works for me (tm), but that's about all I can promise.

Jabber and IRC conversion are untested. For Jabber even _compilation_ is not
tested, as I don't have libpsi handy. The rest of the code all compiles fine
here.

(BTW, do I qualify for the Biggest Commit, Most Changed Files and Longest
Commit Log awards today? :)

CCMAIL: kopete-devel@kde.org

svn path=/trunk/kdenetwork/kopete/; revision=200922