I give up trying to hunt down the uninitialized vars as reported by valgrind. Somewhere in this code they are, but the line number valgrind reports me (an if() statement) surprises me, as at that location each and every member of the TBuddy struct has been set by the call to get().
That said, get() only copies the TBuddy from elsewhere in the TBuddyList and
the initial addition of entries to that list (from OscarProtocol at least)
is severely flawed. OscarProtocol doesn't set most of the fields, which is
fine for QStrings, but not for ints which don't default to 0 if you don't
tell them to.

As a butt ugly hack add a constructor to the struct that explicitly sets
all bools and ints. It works, but is far from clean, obviously.

The best fix to me seems getting rid of TBuddy altogether, though properly
initializing the members might work too.

Either way, the current hack avoids unexpected behaviour and is hence a
suitable fix for now, but also obscures the real problem, so a better fix
is welcome. If one of the people more knowledgeable about the Oscar plugin
could take a look at this, please do. Thanks!

CCMAIL: kopete-devel@kde.org

svn path=/trunk/kdenetwork/kopete/; revision=203410