Two bugfixes here
First bugfix finally enables some sane unloading when Kopete exists. Order is as follows:
1. Protocols unload
2. Each protocols KMM Factory deletes all its KMMs
--- Up to here was already functioning properly AFAIK
3. KMM sends a dying signal to its chatwindow with its widget as a paramater
4. Chatwindow closes this view and deletes it. If all views are deleted then the window
deletes itself

This sequence also ensures that on protocol unload, any chat tabs/windows of that
protocol are removed.

Second bugfix is a tad messy one.. I introduced a single shot timer on the chatview
refresh method to fix the scrolling bug. As suspected, the cause is the contentsHeight()
is not always up to date at the time of the scrollBy() call. This timer delays the call
by 5 ms, with which I have been unable to duplicate the scrolling bug under the same
circumstances it used to be happening.

svn path=/trunk/kdenetwork/kopete/; revision=204744