Tons of stuff: - Vast KMM trimdown. Much of the API was removed, as it is now unneeded. readMessages now gone,  the viewManager handles this type of thing - viewManager flesh out - Implemented old icq style fire and forget window. AKA the "Email" window. Works with all protocols (*MSN, see below)
TODO / ISSUES:
- Finish the email window. It needs a splitter, some better size ajusting, and many of its
slots still need to be implemented. But the basic functionality of it is working.
- For some reason MSN does not work with the email window, because slotSendMessage in the MSN
MM adds msgs to a queue if a chat is not in session or something. Need to investigate
(Gof?)
- Need something in the UI to let users choose which type of window they want to launch when
they click on someone
- setCurrentMessage and currentMessage in the KMM no longer function. Need to investigate if
these are in active use anywhere, and if so, replace them with viewManager->view()->blah
calls

svn path=/trunk/kdenetwork/kopete/; revision=206403