Repeat after me: delete this is evil!
Two other cases where it caused major havoc, this time in the chat windows,
but only in SDI mode.

Took me only 5 hours or so to fully track this down, even with valgrind.
Grmbl. I vote for adding 'delete this' to the list of possibly dangerous
constructs to use, next to printf and setUseShell...

svn path=/trunk/kdenetwork/kopete/; revision=208650