Don't go offline -> unknown -> offline [ -> online ] during connect.
Basically there's no 'unknown' status in MSN. Strictly speaking a contact's
status is unknown when he or she has blocked you, but in that case the user
still appears as offline to you and although there is a dirty trick to find
the difference between being blocked and being really offline I don't think
we'll ever add that to the baseline Kopete MSN plugin.

There's still some support for 'unknown' there, as the protocol does know a
formal 'UNK' status, but AFAIK it's never used. This commit can't hurt anyway.

svn path=/trunk/kdenetwork/kopete/; revision=209123