Remove a lot of code duplication and lay the groundwork for a _LOT_ more: introducing KopeteOnlineStatus. The class is intended to be extremely lightweight, implicitly shared and value-based; the header file does not include ANY other .h files to keep compilation speed fast.
KopeteOnlineStatus encapsulates the local status, the status icon, the
description, and the libkopete-wide status in a single object that can be
passed around everywhere. There's also preliminary support for a caption()
to be used in menus, but that part is incomplete, so you can disregard it
for now (except that you should provide *something* to the constructor).

In the protocols that make heavy use of switch() to convert between internal
status and the Kopete status the gain is already enourmous. This applies to
MSN, Jabber and Gadu. These were also the easiest to port (together with
Oscar, but with only 3 online states the gain is neglectible).

The other plugins all had their issues when porting. WinPopup and SMS seem
not to support the distinction between Unknown and Offline properly and
don't seem to set contact states properly when going online/offline.

Yahoo has its own YahooStatus class that is remarkably similar to
KopeteOnlineStatus, but I only discovered that when porting. With some
effort that class can be ditched easily, but I didn't do that for now.

IRC has some 4 different KopeteContact-like classes. I'd be surprised if
I ported everything correctly there.

ICQ is a total disaster to port because it uses bitmasks instead of ints
or enums.

All in all I expect some regressions and I apologize in advance for those.
I hope to iron those out tomorrow, but I do need reports, as MSN and Oscar
seem to work ok for me. The only problem I have is that the meta contact
icon is not properly updated anymore, but they did last weekend when I
started porting those. I suspect it's one of the recent changes in libkopete
and I'm not sure where it comes from. Will investigate tomorrow.

For now I could already remove three virtuals from all plugins, so creating
a new plugin again has become easier, but there's much more to come. Think
setAway, setAvailable, isOnline, setOnlineStatus and a LOT more stuff that
can now go to libkopete.

CCMAIL: kopete-devel@kde.org

svn path=/trunk/kdenetwork/kopete/; revision=211155