Remplace QString to const QString& in every MNSNotifySocket signal, and more
Sync contactlist for sever only if it is needed (i.e. the contactlist has
been changed by another client) kopete store now the whole contactlist in
the xml file. It also store the version of the server list.

This decrase A LOT the connection time to MSN, especialy with big contact
listes

svn path=/trunk/kdenetwork/kopete/; revision=211777