Next steps towards using KTrader:
- Install the .desktop files in $KDEDIR/share/services so kded can find
  them and trader queries yield proper results
- Add plugin ids to the plugins (at least for now), so we have _something_
  to query
- Make the Kopete/Protocol servicetype inherit Kopete/Plugin
- Port the plugin loader to use trader queries for loading plugins.
  Config file format is unchanged for now, some hacks are in place to
  preserve compatibility with the rather different KTrader API. I will
  be able to remove those when the change is complete.

The listing and querying of plugins (like e.g. the plugin selector does)
is not ported. Ideally we can simply use kpluginselector from kdelibs HEAD
for that and don't even bother porting, but for that I'll almost certainly
need to add stuff to the widget and ask Matthias Kretz for help first :)

svn path=/trunk/kdenetwork/kopete/; revision=213157