Merging ph0nk_work_1 with HEAD.
Compiles.  You can now add multiple AIM accounts.  I'm not sure if it properly imports old contact lists because I dont' ahve an old contact list laying around, so someone please test this.

Problems: Almost all of the group management code needs to be written, like renaming and removing groups.  Support for carrying on conversations with people not in your buddy list is awful right now, and I'm not even really sure if adding new contacts works.

Basically I wanted to merge this in because I want other people to comment on it and give some ideas and test it out.

svn path=/trunk/kdenetwork/kopete/; revision=214442