Assign a unique numeric ID to each metacontact, Used to store metacontact logs.
Add the ability to store logs in a compressed format. Need to add a pref to enable
it.

I also converted the entire to/from XML system to use the QDom everywhere,
resulting
in a noticeable improvement in speed. This indicates a big deal since I don't have
a very large contact list, and if I notice it, users with large lists will have an
even bigger gain I suspect.
----------------------------------------------------------------------

svn path=/trunk/kdenetwork/kopete/; revision=216975