Ran it through indent, and also another code drop.
Not trying to enforce *my* coding style, it should be said, just trying to
get *one* particular uniform style through Jabber. As I'm working on it,
that happens to be mine, but I won't care if that's changed back, just as
long as it stays uniform. The previous code was too ugly to look at.

svn path=/trunk/kdenetwork/kopete/; revision=217559