Load the trackChildNameChange from the xml contactlist. now, the metacontact name will follow the contact name until you give a name yourself. Of course, the default is false, and this will not affect existing contact list.
This has been disabled in the past because we don't have *decent* ui to
re-enable the tracking if you renamed a contact (the current way is to set
an empty name to the contact). But i think that make no sens to don't track
child name if the user has not set an explicit name.

svn path=/trunk/kdenetwork/kopete/; revision=229578