sorry lilachaze, i am reverting your last patch. It cause patch. here is the valgrind backtrace
==29176== Invalid read of size 4
==29176==    at 0x41116433: QDomDocumentPrivate::importNode(QDomNodePrivate cons
t *, bool) (in /usr/lib/qt-3.1.1/lib/libqt-mt.so.3.1.1)
==29176==    by 0x411174C9: QDomDocument::importNode(QDomNode const &, bool) (in
 /usr/lib/qt-3.1.1/lib/libqt-mt.so.3.1.1)
==29176==    by 0x402ADF75: KopeteGroup::toXML(void) (kopetegroup.cpp:102)
==29176==    by 0x4029D69B: KopeteContactList::toXML(void) (kopetecontactlist.cp
p:600)


In many case, i don't beleive the problem is here. The password, the login,
the autoconnect flag ARE NOT pluginData , it is handled dirrectly by the
account.
as i understand, the whole account data is lost, so this is not a plugindata
problem.

CCMAIL:  kde@metafoo.co.uk

svn path=/trunk/kdenetwork/kopete/; revision=234482