CCMAIL: 59541-done@bugs.kde.org
I Had aproximatively the same backtrace when I ported the Action menu to
KXMLGUI.

 #7  0x40811fca in KXMLGUIFactory::removeClient () from /usr/lib/libkdeui.so.4
 #8  0x40811955 in KXMLGUIFactory::addClient () from /usr/lib/libkdeui.so.4

The factory does not allow you to have a client connected to two different UI
so it remove to one in the previous window.  But the previous chatwindow does
not exist anymore.

I hope this patch fix this bug.

svn path=/trunk/kdenetwork/kopete/; revision=234739