Fix the tooltips in KopeteContactListView that were broken since Qt 3.1.
It's a Qt bug for which I committed a patch to qt-copy/patches earlier
today (http://lists.kde.org/?l=kde-cvs&m=105992758126841&w=2). Copying
the short explanation from kopetecontactlistview.cpp:

----
    /*
        The below QTooltip::remove() call seems silly, but is need to work around a Qt bug
        that's present at least in all Qt 3.1.x versions and in 3.2.0. Qt allocates a
        tooltip that covers the entire viewport, which is used by QListView's own tooltips
        if setShowToolTips is set to true (the default) for the tooltips on truncated
        list view items.

        However, this tooltip overrides the custom tooltip that Kopete wants to create
        for itself. As such our own tooltip code is never called and the custom tooltips
        don't appear. To work around the problem we simply remove the existing tooltip
        that QListView creates on the viewport.

        I committed a patch to qt-copy (http://lists.kde.org/?l=kde-cvs&m=105992758126841&w=2),
        but haven't received a reaction from TrollTech yet. Hopefully it will be in upcoming
        Qt versions. In the mean time we check for the presence of the #define that I added
        in the patch.

        Note that Qt 3.0.x is not affected by this bug, but Kopete no longer supports it,
        so there's no #if for that. - Martijn
    */
    setShowToolTips( false );
#ifndef QT_QLISTVIEW_FIXED_TOOLTIPS
    QToolTip::remove( viewport() );
#endif
----

The rest of the commit is basically just a change to use the same approach
that QListView uses for tooltips: create a custom subclassed QToolTip, apply
it to the whole (!) list view and make maybeTip() decide whether or not to
show the tooltip.

Avoids all the flicker that the old approach had and is generally a bit
cleaner.

A special word of thanks applies to Scott Wheeler for helping me in debugging
Qt's intricate internals.

CCMAIL: kopete-devel@kde.org

svn path=/trunk/kdenetwork/kopete/; revision=240775