Start using the new KCDialog framework to replace our crappy ConfigModule.
MSN is ported over, I'll now gradually port the other protocols to use it.
There needs to be a slight change in the way KGenericFactory is used in
Kopete, because its plugin will need its own KInstance, instead of reusing
KGlobal::instance().

I already adapted the KopetePlugin/KopeteProtocol APIs to reflect this, but
for now I just pass KGlobal::instance in all constructors except MSN. Both
the new KCDialog and the new plugin selector will need it later on.

KCDialog is a dialog that dynamically loads KConfigModules (KCModule). This
has the advantage that the preferences are created on demand instead of at
Kopete startup, besides a couple of other nice benefits.

There are a few notes regarding this:
1. There are now two 'Configure Kopete...' entries, one which shows the
   old dialog (without MSN :) and one with the new dialog (and only MSN).
   I don't consider this enough showstopper for HEAD to not commit, it
   will be gone soon enough.

2. I added the entire KCDialog API to libkopete/compat/kcdialog, but I have
   no KDE 3.1 at hand to even try compilation. Some fixes might be needed,
   feel free to commit to the Makefiles.

3. KCDialog is _bleeding_edge_ code, and it has some issues that Matthias
   Kretz is already aware of. Try to stay current with kdelibs/kutils and
   report if necessary, but expect this API to be a rather fast moving
   target in the next couple of days. Apologies for the inconvenience.

CCMAIL: kopete-devel@kde.org

svn path=/trunk/kdenetwork/kopete/; revision=245658