CCMAIL: kopete-devel@kde.org, duncan@kde.org
Yeah, and s/Developer/Credit/ for myself; whoops.

Guys like Nick, are prime candidates for moving to credits as well. And
what's with this stupid Core Developer/Developer split? I'd argue that Ryan
was definitely a core developer ... why not just axe the split? I've always
said it's stupid, and it's even more stupid in this case.

I think the entire credits section needs axing and rewriting.

svn path=/trunk/kdenetwork/kopete/; revision=245863