What smartass decided to fork off KTabWidget from kdelibs HEAD some time ago, include it UNconditionally and even make changes to it?!?
That's bound to cause crashes and symbol clashes with the 'real' KTabWidget
in CVS HEAD. In fact, it did for me.

Renaming to KopeteTabWidget because apparently the two things are not
source compatible.

I wouldn't mind if someone uses the 'real' thing from now on, providing a
copy for KDE 3.1 in libkopete/compat.

svn path=/trunk/kdenetwork/kopete/; revision=246195