Commit the last thing that I definitely wanted to do on n7y: KWallet support for Kopete. This is only enabled for KDE 3.2, as I need fallback code for KConfig in any case.
If the wallet is not available it falls back to the old code.

What's missing (but trivial to add) is the ability to move passwords from
KConfig to KWallet as a migration step and some GUI notification to warn
the user that KWallet cannot be used (or even disable persistent storage
without KWallet). That's for the usability and security departments though :)

svn path=/trunk/kdenetwork/kopete/; revision=246681