First try of exporting a display image. This work fine Kopete <--> Kopete but does not work yet Kopete --> Windows
currently, it export the file in ~/.kde/share/apps/kopete/msnpicture.png
There is an UI (in the editaccount widget) to select it
Anyway, this is not yet account independent

I would like to see it working with the official client, but for some hidden
reason, this won't work. At least, the user will see that you have a picture,
put the picture stay blank. I'll work on that tomorrow

svn path=/trunk/kdenetwork/kopete/; revision=249240