make it compile.
Do we still want to carry this thing with us? It's been disabled from
compile since... well, ages, and i doubt it works if you run it. At the
very least someone will need to port the prefs soon or it won't build in
the future. ConfigModule isn't going to last very long anymore.

svn path=/trunk/kdenetwork/kopete/; revision=249245