- Match the KopeteGroup to the QCheckBoxItem insteads of retreiving it whith the displayname. This will    help if there are two group with the same name, or if there is a psetial group (top-level) - Allow to add empty metacontact... (i don't know hom much this is usefull)
CCMAIL: 64025-done@bugs.kde.org

This is not a fix for the bug 64025 but at least should make the reporter happy.
The 'Default' group was the top-level group in kopete 0.7.0, but now, it is called i18n("Top Level")
if there is still a "Default" group, it's because the name is saved in the contactist.xml or somewere in a server.
That's true when you select it it added a new group, that will not heppends anymore

svn path=/trunk/kdenetwork/kopete/; revision=252464