API for async unloading of the plugins on exit.
Let's start with a big fat note that applies more to kdelibs in general:
---------------------------------------------------------------------------
I had to comment out my call to m_mainWindow->close(). Clicking Quit from
the systray also closes the main window, which derefs kapp. The call to
close in quitKopete calls close() another time, dereffing again, leaving
the refcount at 0 and quitting Kopete.

After adding several kdBacktraces it is really calling it twice, there's
only one KMainWindow at the time and valgrind does NOT (!) spot the use
of deleted memory even though the window has its default W_DestructiveClose
flag set. The only explanation I can come up with is that the destructive
close is posted as a deleteLater() call and still in the event queue.

The really bad thing about this though is that apparently the deref() call
in KMainWindow doesn't check for reentrancy. I wonder how to fix this.
---------------------------------------------------------------------------

There are a couple of rough edges still that I'll address shortly in some
separate smaller commits, but this is the API for the basic plugins.
Protocol plugins will need additional support in KopeteProtocol.

(I don't think we have many non-protocols that can make use of this, but
if there are, feel free to use this API for them already)

svn path=/trunk/kdenetwork/kopete/; revision=256601