Start Kopete Handbook with a 'Getting Started' chapter.
Thanks to Dominik Haumann for helping me get it to build; the credits and licenses chapter was necessary to prevent checkXML from dying from an unknown CDATA at the end of the first chapter.  Any idea why?  I just added a chapter to the existing template.
CCMAIL:kde-doc-english@kde.org

svn path=/trunk/kdenetwork/doc/kopete/; revision=257266