Zack Rusin wrote: > heh, that just made my day. I'm the original author of this class and > don't have the copyright in that file, but a person who committed a two > line patch does. > You go buddy, at this rate we might get your name in most of KDE classes > by next week.
I'm sorry...I'm new to KDE CVS and I just assumed it was common practice,
because Grzegorz asked me to add his name to the copywrite to a file on
the SMS protocol if I committed a very short patch he sent me.  Therefore,
I assumed it was common practice, if not a requirement, to add one's name
to the copywrite when making a change.  But I guess I am wrong.  I meant
no offense - please understand that.  It will not happen again on my part.

svn path=/trunk/kdenetwork/kopete/; revision=257683