I could have spared myself the effort of refactoring KopeteProtocol::slotRefreshStatus, it's old API that's no longer used anymore. Remove it entirely, together with all code that is still using it like KopeteProtocol::status() and the m_status variable.
More important: Use delayed unloading of accounts. It works for me, but I
have only a single account per protocol, and only MSN and Jabber. Anyone
with a more active list of accounts is welcome to test this to death :)

The current approach is to disconnect any account that is still connected
using the async KopeteAccount::disconnect() call. You can safely and slowly
shutdown there, the three second timeout should be plenty, and if it's not
we can increase the timeout a bit. No need for hacks to kill sockets and
shutdown in a hurry...

When all accounts are disconnected we will emit readyForUnload and actually
unload the protocol.

svn path=/trunk/kdenetwork/kopete/; revision=258124