Save the account on the disk before they are removed
the problem is that accounts was not designed to be saved in a rc file, but in a xml file.  we could realy optimize the way accounts are saved after KDE 3.2.
Actualy, it make no sens anymore to have them a a KopetePluginDataObject which was an helper for writing on the XML file.

CCMAIL: klingens@kde.org

svn path=/trunk/kdenetwork/kopete/; revision=258289