Optimization of the starting.
Finaly, i commited the patch i submit to the mailing list.

This patch built the main window and the contactlist UI  _after_ the loading of the contactlist from the xml file, and the creation of every accounts and contacts.  The effect is that every changes when contacts are added or initial status set are not yet catched by the UI at start.

svn path=/trunk/kdenetwork/kopete/; revision=259724