myself() is no longer pure virtual, instead the plugins can call setMyself() in KopeteAccount now.
This cuts a bit of code duplication in the protocols (the myself() implementation
in the current codebase is the finally same for all protocols unlike in the
past when we had to use the virtual)

The other reason for this commit (besides cleanup) is that it allows SAFE
access to myself() before it's initialized, it's guaranteed to return a
NULL pointer in that case instead of uninitialized memory.

I'll commit the plugins separately when my rebuild there is done (MSN is
tested and works, but has to wait for the others ;), because I have the
hope that the diff now still fits in the mail sent by kde-cvs.

CCMAIL: kopete-devel@kde.org

svn path=/trunk/kdenetwork/kopete/; revision=259946