Ok, it took me an afternoon and evening of toying with KCacheGrind because my contact list is a bit special, but now I know where all the slowness that Olivier complained about during Kopete startup goes.
For every change to a KopeteMetaContact in a KopeteGroup (!!!) the LVI will
call KopeteGroupLVI::refreshDisplayName() 6 times during Kopete startup only,
not counting actually going online and any subsequent activity on the meta
contact during the Kopete runtime.

Although these initial 6 calls are quite excessive they hardly hurt, save
for one thing: they force a sort() on the entire list view every time.

I have 20 meta contacts in top level and these ate only 1/4 of the CPU time
eaten by 4 (!) meta contacts that are in a group, not even counting the
other code paths that triggered refreshDisplayName().

If you start Kopete in cachegrind with a single plugin and quit it using
the GUI when the plugin is unloaded about 4.5% of execution time goes to
sorting the list view.

Added a delayed sort that waits half a second before sorting. For me this
combines all 24 sort requests into a single, making the 4.5% something
neglectible now.

I'm glad I did this, because this change (unlike Olivier's hack to delay
the main window instantiation) also affects Kopete's normal run time. It
should make it respond a fair bit quicker when going online/offline.

Reducing the 6 calls per KMC on startup would be another solution, but
doesn't avoid the problem in other places, so this optimization is needed
regardless.

The one remaining question is why toplevel items are not sorted. As long as
there's at least one item, in a group the sort is called, but with everything
toplevel it's not. Olivier?

CCMAIL: kopete-devel@kde.org

svn path=/trunk/kdenetwork/kopete/; revision=260028